Move getLocalDataDir to util.ts
